requirements.txt
streamlit
langchain
langchain-groq
langgraph
langchain_mcp_adapters
ffmpeg-python
moviepy

